refactor: move performance/health/logging into telemetry/ (Plan E)
