+++ src/mcp/tools.ts
@@ -12,10 +12,10 @@ import { z } from "zod";
 import { remoteClient } from "../activitypub/remote-client.js";
 import { dynamicInstanceDiscovery } from "../discovery/dynamic-instance-discovery.js";
 import { instanceDiscovery } from "../discovery/instance-discovery.js";
-import { healthChecker } from "../health-check.js";
-import { performanceMonitor } from "../performance-monitor.js";
 import { validateActorIdentifier, validateDomain, validateQuery } from "../server/index.js";
 import type { RateLimiter } from "../server/rate-limiter.js";
+import { healthChecker } from "../telemetry/health-check.js";
+import { performanceMonitor } from "../telemetry/performance-monitor.js";
 import { formatErrorWithSuggestion, getErrorMessage, stripHtmlTags } from "../utils.js";
 import { ActorIdentifierSchema, DomainSchema, QuerySchema } from "../validation/schemas.js";
 import { trackedMcpServer } from "./capabilities.js";
