+++ src/telemetry/performance-monitor.ts
