+++ tests/unit/mcp-tools-export.test.ts
@@ -59,7 +59,7 @@ vi.mock("../../src/activitypub/remote-client.js", () => ({
   },
 }));
 
-vi.mock("../../src/performance-monitor.js", () => ({
+vi.mock("../../src/telemetry/performance-monitor.js", () => ({
   performanceMonitor: {
     startRequest: vi.fn().mockReturnValue("req-123"),
     endRequest: vi.fn(),
