+++ tests/unit/performance-monitor.test.ts
@@ -28,7 +28,7 @@ describe("PerformanceMonitor", () => {
 
   describe("singleton instance", () => {
     it("should export performanceMonitor singleton", async () => {
-      const { performanceMonitor } = await import("../../src/performance-monitor.js");
+      const { performanceMonitor } = await import("../../src/telemetry/performance-monitor.js");
       expect(performanceMonitor).toBeDefined();
       expect(typeof performanceMonitor.startRequest).toBe("function");
       expect(typeof performanceMonitor.endRequest).toBe("function");
@@ -36,7 +36,7 @@ describe("PerformanceMonitor", () => {
     });
 
     it("should have getMetrics returning PerformanceMetrics", async () => {
-      const { performanceMonitor } = await import("../../src/performance-monitor.js");
+      const { performanceMonitor } = await import("../../src/telemetry/performance-monitor.js");
       const metrics = performanceMonitor.getMetrics();
 
       expect(metrics).toHaveProperty("requestCount");
@@ -51,7 +51,7 @@ describe("PerformanceMonitor", () => {
     });
 
     it("should have getHealthStatus returning health data", async () => {
-      const { performanceMonitor } = await import("../../src/performance-monitor.js");
+      const { performanceMonitor } = await import("../../src/telemetry/performance-monitor.js");
       const health = performanceMonitor.getHealthStatus();
 
       expect(health).toHaveProperty("status");
@@ -61,14 +61,14 @@ describe("PerformanceMonitor", () => {
     });
 
     it("should have getRequestHistory returning array", async () => {
-      const { performanceMonitor } = await import("../../src/performance-monitor.js");
+      const { performanceMonitor } = await import("../../src/telemetry/performance-monitor.js");
       const history = performanceMonitor.getRequestHistory();
 
       expect(Array.isArray(history)).toBe(true);
     });
 
     it("should have getOperationMetrics returning metrics object", async () => {
-      const { performanceMonitor } = await import("../../src/performance-monitor.js");
+      const { performanceMonitor } = await import("../../src/telemetry/performance-monitor.js");
       const opMetrics = performanceMonitor.getOperationMetrics("test-operation");
 
       expect(opMetrics).toHaveProperty("count");
@@ -79,7 +79,7 @@ describe("PerformanceMonitor", () => {
     });
 
     it("should be able to stop without error", async () => {
-      const { performanceMonitor } = await import("../../src/performance-monitor.js");
+      const { performanceMonitor } = await import("../../src/telemetry/performance-monitor.js");
       expect(() => performanceMonitor.stop()).not.toThrow();
     });
   });
@@ -87,7 +87,7 @@ describe("PerformanceMonitor", () => {
 
 describe("PerformanceMetrics types", () => {
   it("should export PerformanceMetrics interface", async () => {
-    const module = await import("../../src/performance-monitor.js");
+    const module = await import("../../src/telemetry/performance-monitor.js");
     expect(module.performanceMonitor).toBeDefined();
     const metrics = module.performanceMonitor.getMetrics();
 
@@ -112,7 +112,7 @@ describe("PerformanceMonitor.startMetricsCollection unref (H5)", () => {
   });
 
   it("unrefs the metrics interval so it does not keep the event loop alive", async () => {
-    const { performanceMonitor } = await import("../../src/performance-monitor.js");
+    const { performanceMonitor } = await import("../../src/telemetry/performance-monitor.js");
     // The constructor already called startMetricsCollection() because METRICS_ENABLED=true
     const interval = (performanceMonitor as unknown as { metricsInterval?: NodeJS.Timeout })
       .metricsInterval;
@@ -144,7 +144,7 @@ describe("PerformanceMonitor with metrics enabled", () => {
   });
 
   it("should track request when metrics enabled", async () => {
-    const { performanceMonitor } = await import("../../src/performance-monitor.js");
+    const { performanceMonitor } = await import("../../src/telemetry/performance-monitor.js");
 
     const requestId = performanceMonitor.startRequest("test-op", { key: "value" });
     expect(requestId).not.toBe("");
@@ -155,7 +155,7 @@ describe("PerformanceMonitor with metrics enabled", () => {
   });
 
   it("should track failed request", async () => {
-    const { performanceMonitor } = await import("../../src/performance-monitor.js");
+    const { performanceMonitor } = await import("../../src/telemetry/performance-monitor.js");
 
     const requestId = performanceMonitor.startRequest("failing-op");
     performanceMonitor.endRequest(requestId, false, "Test error message");
@@ -167,7 +167,7 @@ describe("PerformanceMonitor with metrics enabled", () => {
   });
 
   it("should handle unknown requestId gracefully", async () => {
-    const { performanceMonitor } = await import("../../src/performance-monitor.js");
+    const { performanceMonitor } = await import("../../src/telemetry/performance-monitor.js");
 
     expect(() => performanceMonitor.endRequest("unknown-request-id", true)).not.toThrow();
 
@@ -175,7 +175,7 @@ describe("PerformanceMonitor with metrics enabled", () => {
   });
 
   it("should handle empty requestId gracefully", async () => {
-    const { performanceMonitor } = await import("../../src/performance-monitor.js");
+    const { performanceMonitor } = await import("../../src/telemetry/performance-monitor.js");
 
     expect(() => performanceMonitor.endRequest("", true)).not.toThrow();
 
@@ -183,7 +183,7 @@ describe("PerformanceMonitor with metrics enabled", () => {
   });
 
   it("should handle multiple requests and return metrics", async () => {
-    const { performanceMonitor } = await import("../../src/performance-monitor.js");
+    const { performanceMonitor } = await import("../../src/telemetry/performance-monitor.js");
 
     const id1 = performanceMonitor.startRequest("multi-test-1");
     performanceMonitor.endRequest(id1, true);
@@ -199,7 +199,7 @@ describe("PerformanceMonitor with metrics enabled", () => {
   });
 
   it("should maintain history size limit", async () => {
-    const { performanceMonitor } = await import("../../src/performance-monitor.js");
+    const { performanceMonitor } = await import("../../src/telemetry/performance-monitor.js");
 
     // Create many requests
     for (let i = 0; i < 100; i++) {
@@ -214,7 +214,7 @@ describe("PerformanceMonitor with metrics enabled", () => {
   });
 
   it("should get operation-specific metrics", async () => {
-    const { performanceMonitor } = await import("../../src/performance-monitor.js");
+    const { performanceMonitor } = await import("../../src/telemetry/performance-monitor.js");
 
     const id1 = performanceMonitor.startRequest("specific-op");
     performanceMonitor.endRequest(id1, true);
